requirements.txt
PyYAML==3.13
prettytable==0.7.2 
pytest==4.0.0 
